requirements.txt
absl-py==1.2.0
astunparse==1.6.3
attrs==22.1.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
comtypes==1.1.14
cycler==0.11.0
face-recognition-models==0.3.0
flatbuffers==1.12
fonttools==4.33.3
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
gTTS==2.2.4
h5py==3.7.0
idna==2.10
imutils==0.5.4
Jinja2==3.0.1
joblib==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.3
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
mediapipe==0.8.10.1
numpy==1.23.0
oauthlib==3.2.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
Pillow==8.4.0
playsound==1.2.2
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.1.2
pyparsing==3.0.9
pypiwin32==223
pytesseract==0.3.10
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2022.1
pywin32==304
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.8.1
six==1.16.0
sklearn==0.0
soupsieve==2.3.2.post1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
tesseract==0.1.3
threadpoolctl==3.1.0
tkdesigner==1.0.4
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
